+++ src/components/Action/Action.test.tsx
@@ -127,10 +127,8 @@ describe('formulaire d‘ajout d‘une action', () => {
       expect(pluriannuelle).not.toBeChecked()
       const selecteurAnneeDeDebut = within(formulaire).getByLabelText('Année de début de l‘action')
       expect(selecteurAnneeDeDebut).toHaveAttribute('name', 'anneeDeDebut')
-      expect(selecteurAnneeDeDebut.tagName).toBe('SELECT')
       const selecteurAnneeDeFin = within(formulaire).getByLabelText('Année de fin de l‘action')
       expect(selecteurAnneeDeFin).toHaveAttribute('name', 'anneeDeFin')
-      expect(selecteurAnneeDeFin.tagName).toBe('SELECT')
       expect(selecteurAnneeDeFin).toBeDisabled()
       const titreSectionInformationBudget = within(formulaire).getByText('Information sur le budget et le financement', { selector: 'p' })
       expect(titreSectionInformationBudget).toBeInTheDocument()
@@ -241,8 +239,8 @@ describe('formulaire d‘ajout d‘une action', () => {
       afficherFormulaireDeCreationAction()
 
       // WHEN
-      jouvreLeDrawerDeCoFinancementFormulaireAjout()
-      const drawer = screen.getByRole('dialog', { name: 'Ajouter un co-financement' })
+      jOuvreLeFormulairePourAjouterUnCoFinancement()
+      const drawer = screen.getByRole('dialog', { hidden: false, name: 'Ajouter un co-financement' })
       jeCreeUnCofinancementDansLeDrawer(drawer)
       const boutonEnregistrer = within(drawer).getByRole('button', { name: 'Enregistrer' })
       fireEvent.click(boutonEnregistrer)
@@ -405,110 +403,9 @@ describe('formulaire d‘ajout d‘une action', () => {
       expect(screen.getByLabelText('Année de fin de l‘action')).toBeDisabled()
     })
   })
-
-  describe('drawer d‘ajout d‘un co-financement', () => {
-    it('étant un utilisateur, lorsque je clique sur le bouton ajouter un financement, alors le drawer s‘ouvre', () => {
-      // GIVEN
-      afficherFormulaireDeCreationAction()
-
-      // WHEN
-      jouvreLeDrawerDeCoFinancementFormulaireAjout()
-
-      // THEN
-      const drawer = screen.getByRole('dialog', { name: 'Ajouter un co-financement' })
-      expect(drawer).toBeInTheDocument()
-      const titre = within(drawer).getByRole('heading', { level: 1, name: 'Ajouter un co-financement' })
-      expect(titre).toBeInTheDocument()
-      const boutonFermer = within(drawer).getByRole('button', { name: 'Fermer' })
-      expect(boutonFermer).toBeInTheDocument()
-      const texteDInstruction = within(drawer).getByText('Précisez l‘origine du financement', { selector: 'p' })
-      expect(texteDInstruction).toBeInTheDocument()
-      const selecteurOrigineDuFinancement = within(drawer).getByLabelText('Membre de la gouvernance')
-      expect(selecteurOrigineDuFinancement.tagName).toBe('SELECT')
-      expect(selecteurOrigineDuFinancement).toHaveAttribute('name', 'cofinanceur')
-      const montantDuFinancement = within(drawer).getByLabelText('Montant du financement *')
-      expect(montantDuFinancement).toBeRequired()
-      expect(montantDuFinancement).toHaveAttribute('type', 'number')
-      expect(montantDuFinancement).toHaveAttribute('min', '0')
-      expect(montantDuFinancement).toHaveAttribute('placeholder', '5 000')
-      expect(montantDuFinancement).toBeRequired()
-      const boutonEnregistrer = within(drawer).getByRole('button', { name: 'Enregistrer' })
-      expect(boutonEnregistrer).toBeDisabled()
-    })
-
-    it('étant un utilisateur, lorsque je remplis correctement le formulaire d‘ajout d‘un co-financement, alors il est ajouté', async () => {
-      // GIVEN
-      afficherFormulaireDeCreationAction()
-
-      // WHEN
-      jouvreLeDrawerDeCoFinancementFormulaireAjout()
-      const drawer = screen.getByRole('dialog', { name: 'Ajouter un co-financement' })
-      jeCreeUnCofinancementDansLeDrawer(drawer)
-      const boutonEnregistrer = within(drawer).getByRole('button', { name: 'Enregistrer' })
-      fireEvent.submit(boutonEnregistrer)
-
-      // THEN
-      const formulaire = screen.getByRole('form', { name: 'Ajouter une action à la feuille de route' })
-      const listeCofinancements = await within(formulaire).findAllByRole('listitem')
-      expect(listeCofinancements).toHaveLength(1)
-      const premierCofinancement = within(listeCofinancements[0]).getByText('CC des Monts du Lyonnais')
-      expect(premierCofinancement).toBeInTheDocument()
-      const montantPremierCofinancement = within(listeCofinancements[0]).getByText('1000 €')
-      expect(montantPremierCofinancement).toBeInTheDocument()
-    })
-
-    it('étant un utilisateur, lorsque je remplis correctement le formulaire de modification d‘un co-financement, alors il est ajouté', async () => {
-      // GIVEN
-      afficherFormulaireDeModificationAction()
-
-      // WHEN
-      jouvreLeDrawerDeCoFinancementFormulaireModification()
-      const drawer = screen.getByRole('dialog', { name: 'Ajouter un co-financement' })
-      jeCreeUnCofinancementDansLeDrawer(drawer)
-      const boutonEnregistrer = within(drawer).getByRole('button', { name: 'Enregistrer' })
-      fireEvent.submit(boutonEnregistrer)
-
-      // THEN
-      const formulaire = screen.getByRole('form', { name: 'Modifier une action' })
-      const listeCofinancements = await within(formulaire).findAllByRole('listitem')
-      expect(listeCofinancements).toHaveLength(5)
-      const cinquiemeCofinancement = within(listeCofinancements[4]).getByText('CC des Monts du Lyonnais')
-      expect(cinquiemeCofinancement).toBeInTheDocument()
-      const montantCinquiemeCofinancement = within(listeCofinancements[4]).getByText('1000 €')
-      expect(montantCinquiemeCofinancement).toBeInTheDocument()
-    })
-
-    it('lorque je clique sur le bouton fermer du drawer d‘ajout d‘un co-financement, dans le formulaire d‘ajout d‘une action, alors le drawer se ferme', () => {
-      // GIVEN
-      afficherFormulaireDeCreationAction()
-
-      // WHEN
-      jouvreLeDrawerDeCoFinancementFormulaireAjout()
-      const drawer = screen.getByRole('dialog', { name: 'Ajouter un co-financement' })
-      const boutonFermer = within(drawer).getByRole('button', { name: 'Fermer' })
-      fireEvent.click(boutonFermer)
-
-      // THEN
-      expect(drawer).not.toBeVisible()
-    })
-
-    it('lorque je clique sur le bouton fermer du drawer d‘ajout d‘un co-financement, dans le formulaire de modification d‘une action, alors le drawer se ferme', () => {
-      // GIVEN
-      afficherFormulaireDeModificationAction()
-
-      // WHEN
-      jouvreLeDrawerDeCoFinancementFormulaireModification()
-      const drawer = screen.getByRole('dialog', { name: 'Ajouter un co-financement' })
-      const boutonFermer = within(drawer).getByRole('button', { name: 'Fermer' })
-      fireEvent.click(boutonFermer)
-
-      // THEN
-      expect(drawer).not.toBeVisible()
-    })
-  })
 })
 
-function afficherFormulaireDeCreationAction(options?: Partial<Parameters<typeof renderComponent>[1]>): void {
+export function afficherFormulaireDeCreationAction(options?: Partial<Parameters<typeof renderComponent>[1]>): void {
   renderComponent(
     <AjouterUneAction
       action={actionVideViewModelFactory()}
@@ -541,7 +438,7 @@ function afficherFormulaireDeCreationAction(options?: Partial<Parameters<typeof 
   )
 }
 
-function afficherFormulaireDeModificationAction(modifierUneActionAction: Mock = vi.fn()): void {
+export function afficherFormulaireDeModificationAction(modifierUneActionAction: Mock = vi.fn()): void {
   renderComponent(
     <ModifierUneAction
       action={actionViewModelFactory()}
@@ -651,20 +548,15 @@ function jeTapeLaDescriptionDeLaction(): HTMLElement {
   return description
 }
 
-function jouvreLeDrawerDeCoFinancementFormulaireAjout(): void {
+export function jOuvreLeFormulairePourAjouterUnCoFinancement(): void {
   const formulaire = screen.getByRole('form', { name: 'Ajouter une action à la feuille de route' })
   const boutonAjouterUnCoFinanacement = within(formulaire).getByRole('button', { name: 'Ajouter un financement' })
   fireEvent.click(boutonAjouterUnCoFinanacement)
 }
-function jouvreLeDrawerDeCoFinancementFormulaireModification(): void {
-  const formulaire = screen.getByRole('form', { name: 'Modifier une action' })
-  const boutonAjouterUnCoFinanacement = within(formulaire).getByRole('button', { name: 'Ajouter un financement' })
-  fireEvent.click(boutonAjouterUnCoFinanacement)
-}
 
-function jeCreeUnCofinancementDansLeDrawer(drawer: HTMLElement): void {
-  const selecteurOrigineDuFinancement = within(drawer).getByLabelText('Membre de la gouvernance')
+export function jeCreeUnCofinancementDansLeDrawer(drawer: HTMLElement): void {
+  const selecteurOrigineDuFinancement = within(drawer).getByRole('combobox', { name: 'Membre de la gouvernance' })
   fireEvent.change(selecteurOrigineDuFinancement, { target: { value: 'CC des Monts du Lyonnais' } })
-  const montantDuFinancement = within(drawer).getByLabelText('Montant du financement *')
+  const montantDuFinancement = within(drawer).getByRole('spinbutton', { name: /Montant du financement \*/ })
   fireEvent.change(montantDuFinancement, { target: { value: 1000 } })
 }
